requirements.txt
bitso-py==3.0.1
requests==2.24.0
tauros-api==1.5.1
python-dotenv==0.14.0
simplejson==3.17.2
black==22.6.0
websocket-client==1.3.3
